pyproject.toml
[project]
name = "single-auth-microservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "celery>=5.6.2",
    "fastapi[all,standard]>=0.129.0",
    "polars>=1.38.1",
    "pydantic-settings>=2.13.1",
    "pyodbc>=5.3.0",
    "python-jose[cryptography]>=3.5.0",
    "ruff>=0.15.1",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.17",
]
